Refactored SaveButton into its own component
